Convert all tasks to use REAL indices

+++ py/simple_process_pdfs.py
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS documents (
     filename TEXT,
     content TEXT,
     keywords TEXT,
-    faiss_index INTEGER
+    faiss_index REAL
 )
 ''')
 
@@ -69,7 +69,7 @@ faiss_index = faiss.IndexFlatL2(d)
 # Add vectors to FAISS and update SQLite with FAISS indices
 faiss_index.add(embeddings)
 for i, (doc_id, _) in enumerate(documents):
-    cursor.execute('UPDATE documents SET faiss_index = ? WHERE id = ?', (int(i), doc_id))
+    cursor.execute('UPDATE documents SET faiss_index = ? WHERE id = ?', (float(i), doc_id))
 conn.commit()
 
 # Save the FAISS index to file
